+++ src/components/SavedQueries/ListSavedQueries.js
@@ -7,7 +7,6 @@ import { useNamespace, CalloutContext } from '@folio/stripes/core';
 import { Paneset, Pane, MultiColumnList, IconButton, ConfirmationModal } from '@folio/stripes/components';
 
 
-// eslint-disable-next-line no-unused-vars
 function ListSavedQueries({ queries, deleteQuery }) {
   const [queryToDelete, setQueryToDelete] = useState();
   const history = useHistory();
@@ -15,9 +14,9 @@ function ListSavedQueries({ queries, deleteQuery }) {
   const namespace = getNamespace({ key: 'formState' });
   const callout = useContext(CalloutContext);
 
-  const executeQuery = (_unusedEvent, item) => {
-    localforage.setItem(namespace, { tables: item.json.tables })
-      .then(() => history.push(`/ldp${item.autoRun ? '?execute' : ''}`));
+  const selectQuery = async (_unusedEvent, item) => {
+    await localforage.setItem(namespace, { META: item, tables: item.json.tables });
+    history.push(`/ldp${item.autoRun ? '?execute' : ''}`);
   };
 
   function maybeDeleteQuery(e, item) {
@@ -72,7 +71,7 @@ function ListSavedQueries({ queries, deleteQuery }) {
             created: r => new Date(r.created).toLocaleString(),
             deleteQuery: r => <IconButton icon="trash" onClick={e => maybeDeleteQuery(e, r)} />
           }}
-          onRowClick={executeQuery}
+          onRowClick={selectQuery}
         />
 
         {queryToDelete !== undefined &&